Switch from interfaces to generic structs
Switch from interfaces to generic structs to support the correct semantics on `Dispose`.

At this point there are two design decisions that will have to be made:

1. How to support regular `IDisposable`
    - Extension method?
    - Implicit cast operator?
    - Explicit cast operator?
2. If the borrowed version should be "strict" and throw when attempting
to dispose. It would make the code "safer" at the cost of verbosity when
using regular `IDisposable`. Currently the constructor takes a `bool`
and supports both behaviors.