Rename to "Lagan" and create solution
Lagan (from Wikipedia):
Also called "ligan", the term refers to goods that are cast overboard
and are heavy enough to sink to the ocean floor, but are tied or
otherwise linked to a floating marker, such as a buoy or cork, so
that they can be found again by whoever marked the item. Lagan can
also refer to large objects that are trapped within the sinking vessel.

According to maritime law, a buoy, or otherwise floating object,
constitutes as sufficient grounds in laying claim to an artifact.

**In other words, when faced with a lagan, one is required to return
the artifact to its rightful owner.** (emphasis mine)

Also, create a solution file and update namespaces and assembly names.